final-project/final-deploy/opt/code/requirements.txt
pandas==1.1.5
numpy==1.19.3
tensorflow==2.5.0
flask==2.0.1
gunicorn==20.1.0
pillow==8.2.0